+++ package.json
@@ -44,15 +44,15 @@
     "@hookform/devtools": "^4.3.1",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "@storybook/addon-actions": "^8.1.11",
-    "@storybook/addon-essentials": "^8.1.11",
-    "@storybook/addon-interactions": "^8.1.11",
-    "@storybook/addon-links": "^8.1.11",
-    "@storybook/addon-themes": "^8.1.11",
-    "@storybook/blocks": "^8.1.11",
-    "@storybook/preview-api": "^8.1.11",
-    "@storybook/react": "^8.1.11",
-    "@storybook/react-vite": "^8.1.11",
+    "@storybook/addon-actions": "^8.2.5",
+    "@storybook/addon-essentials": "^8.2.5",
+    "@storybook/addon-interactions": "^8.2.5",
+    "@storybook/addon-links": "^8.2.5",
+    "@storybook/addon-themes": "^8.2.5",
+    "@storybook/blocks": "^8.2.5",
+    "@storybook/preview-api": "^8.2.5",
+    "@storybook/react": "^8.2.5",
+    "@storybook/react-vite": "^8.2.5",
     "@storybook/testing-library": "^0.2.2",
     "@testing-library/dom": "^9.3.4",
     "@testing-library/jest-dom": "^6.4.2",
@@ -94,7 +94,7 @@
     "react-click-away-listener": "^2.2.3",
     "react-dom": "^18.2.0",
     "semantic-release": "^23.0.2",
-    "storybook": "^8.1.11",
+    "storybook": "^8.2.5",
     "storybook-addon-pseudo-states": "^3.0.0",
     "tailwind-merge": "^2.2.1",
     "tailwindcss": "^3.4.1",
