+++ src/connect/constants/options.tsx
@@ -1,4 +1,12 @@
-import { CLOUD, LOCAL, NodeOption, PUBLIC } from '@/connect';
+import {
+    CLOUD,
+    DebugNodeOption,
+    LOCAL,
+    NodeOption,
+    NormalNodeOption,
+    OptionMetadata,
+    PUBLIC,
+} from '@/connect';
 import { Icon } from '@/powerhouse';
 
 export const NEW = 'NEW';
@@ -81,4 +89,48 @@ export const locationInfoByLocation = {
         description: 'Public and available to everyone.',
         icon: <Icon name="drive" />,
     },
+} as const;
+
+export const debugNodeOptionsMap: Record<DebugNodeOption, OptionMetadata> = {
+    [ADD_TRIGGER]: {
+        label: 'Add Trigger',
+        icon: <Icon name="plus" className="text-orange-900" />,
+    },
+    [REMOVE_TRIGGER]: {
+        label: 'Remove Trigger',
+        icon: <Icon name="xmark" className="text-orange-900" />,
+    },
+    [ADD_INVALID_TRIGGER]: {
+        label: 'Add Trigger',
+        icon: <Icon name="exclamation" className="text-orange-900" />,
+    },
+} as const;
+
+export const normalNodeOptionsMap: Record<NormalNodeOption, OptionMetadata> = {
+    [DUPLICATE]: {
+        label: 'Duplicate',
+        icon: <Icon name="files-earmark" />,
+    },
+    [NEW_FOLDER]: {
+        label: 'New Folder',
+        icon: <Icon name="folder-plus" />,
+    },
+    [RENAME]: {
+        label: 'Rename',
+        icon: <Icon name="pencil" />,
+    },
+    [DELETE]: {
+        label: 'Delete',
+        icon: <Icon name="trash" />,
+        className: 'text-red-900',
+    },
+    [SETTINGS]: {
+        label: 'Settings',
+        icon: <Icon name="gear" />,
+    },
+} as const;
+
+export const nodeOptionsMap: Record<NodeOption, OptionMetadata> = {
+    ...debugNodeOptionsMap,
+    ...normalNodeOptionsMap,
 } as const;
