+++ src/connect/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './address';
 export * from './get-folder-status';
 export * from './mocks';
-export * from './node-options';
